refactor: GraphDataConvertor 위치 변경
GraphDataConvertor는 data 패키지와 Graph 패키지 간 데이터 전환을 담당하므로 두 패키지와 같은 레벨에 있도록 변경
