+++ android/app/src/main/java/app/priceguard/ui/home/list/ProductListViewModel.kt
@@ -2,8 +2,8 @@ package app.priceguard.ui.home.list
 
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
+import app.priceguard.data.GraphDataConverter
 import app.priceguard.data.dto.ProductErrorState
-import app.priceguard.data.graph.GraphDataConverter
 import app.priceguard.data.network.ProductRepositoryResult
 import app.priceguard.data.repository.ProductRepository
 import app.priceguard.materialchart.data.GraphMode
